+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.5.0-rc.5] - 2023-03-30
+### Added
+- `@rarimo/nft-checkout`: Selectable token to swap
+- `@rarimo/nft-checkout`: Disabled native token and wrrapped native tokens
+
+### Fixed
+- `@rarimo/nft-checkout`: Automatic change chain
+- `@rarimo/provider`: Automatic change chain
+- `@rarimo/provider`: Load supported tokens
 
 ## [1.5.0-rc.4] - 2023-03-24
 ### Fixed
@@ -124,7 +133,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
-[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.4...HEAD
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.5...HEAD
+[1.5.0-rc.5]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.4...1.5.0-rc.5
 [1.5.0-rc.4]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.3...1.5.0-rc.4
 [1.5.0-rc.3]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.2...1.5.0-rc.3
 [1.5.0-rc.2]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.1...1.5.0-rc.2